@@ -1,6 +1,6 @@
 {
   "name": "svelte-breadcrumbs",
-  "version": "1.2.8",
+  "version": "2.0.0",
   "scripts": {
     "dev": "vite dev",
     "build": "vite build && npm run package",
@@ -22,18 +22,19 @@
     "!dist/**/*.spec.*"
   ],
   "peerDependencies": {
-    "svelte": "3.x || 4.x"
+    "svelte": "5.x"
   },
   "devDependencies": {
-    "@sveltejs/kit": "1.24.0",
-    "@sveltejs/adapter-auto": "2.1.0",
-    "@sveltejs/package": "2.2.2",
-    "publint": "0.1.16",
-    "svelte": "4.2.0",
-    "svelte-check": "3.5.1",
-    "tslib": "2.6.2",
-    "typescript": "5.2.2",
-    "vite": "4.4.9"
+    "@sveltejs/adapter-auto": "^3.3.1",
+    "@sveltejs/kit": "^2.7.3",
+    "@sveltejs/package": "^2.3.7",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "publint": "^0.2.12",
+    "svelte": "^5.1.6",
+    "svelte-check": "^4.0.5",
+    "tslib": "^2.8.0",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.10"
   },
   "svelte": "./dist/index.js",
   "types": "./dist/index.d.ts",
