@@ -2,9 +2,15 @@
   import { page } from "$app/stores";
   import type { Crumb } from "$lib";
   import Breadcrumbs from "$lib/components/Breadcrumbs.svelte";
+  import type { Snippet } from "svelte";
   import type { MyCrumbMetadata } from "./types";
+  interface Props {
+    children?: Snippet;
+  }
 
-  $: pageDataCrumbs = $page.data.crumbs as Crumb<MyCrumbMetadata>[] | undefined;
+  let { children }: Props = $props();
+
+  let pageDataCrumbs = $derived($page.data.crumbs as Crumb<MyCrumbMetadata>[] | undefined);
 </script>
 
 <Breadcrumbs
@@ -12,24 +18,26 @@
   routeId={$page.route.id}
   pageData={$page.data}
   crumbs={pageDataCrumbs}
-  let:crumbs
+  
 >
-  <div>
-    <span><a href="/">Home</a></span>
-    {#each crumbs as c}
-      <span>/</span>
-      <span>
-        <a href={c.url}>
-          <!-- 
-          Pass in the glob import of the route svelte modules as well as
-          any data the routes can use to try to fill in any info.
-          -->
-          {c.title}
-          {c.metadata ? `(${c.metadata.extraValue})` : ""}
-        </a>
-      </span>
-    {/each}
-  </div>
+  {#snippet children({ crumbs })}
+    <div>
+      <span><a href="/">Home</a></span>
+      {#each crumbs as c}
+        <span>/</span>
+        <span>
+          <a href={c.url}>
+            <!-- 
+            Pass in the glob import of the route svelte modules as well as
+            any data the routes can use to try to fill in any info.
+            -->
+            {c.title}
+            {c.metadata ? `(${c.metadata.extraValue})` : ""}
+          </a>
+        </span>
+      {/each}
+    </div>
+  {/snippet}
 </Breadcrumbs>
 
-<slot />
+{@render children?.()}
