@@ -1,4 +1,4 @@
-<script context="module" lang="ts">
+<script module lang="ts">
   // You can set the page title with a function that will be given
   // data from the component in the layout
   export function getPageTitle(pageData: any) {
